[patch 2/5] ras/rastools: Consolidate sub tests
Consolidate sub tests to ensure all tools belonging to a particular
package are part of single test(py file)
